Modularized install script incl. help; Segmented out MetricRunner API
