+++ frontend/js/helpers/config.js.example
@@ -2,9 +2,9 @@
 API_URL = "__API_URL__"
 METRICS_URL = "__METRICS_URL__"
 
-ACTIVATE_METRIC_RUNNER = true; // MetricRunner is always active bc open source. But can be deactivated if not needed
+ACTIVATE_METRIC_RUNNER = __ACTIVATE_METRIC_RUNNER__;
 ACTIVATE_CARBON_DB = __ACTIVATE_CARBON_DB__;
-ACTIVATE_ECO_CI = true; // Eco CI is always active bc open source. But can be deactivated if not needed
+ACTIVATE_ECO_CI = __ACTIVATE_ECO_CI__;
 ACTIVATE_POWER_HOG = __ACTIVATE_POWER_HOG__;
